% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmbetabinom.R
\name{dcbetabinom}
\alias{dcbetabinom}
\title{Contaminated Mean-Parameterized Beta-Binomial Density}
\usage{
dcbetabinom(x, size, mu, sigma, delta, eta, log = FALSE)
}
\arguments{
\item{x}{Vector of non-negative integer quantities.}

\item{size}{Number (or vector of length x) of trials (non-negative integer).}

\item{mu}{Analogue of the binomial probability of success parameter, where 0 < mu < 1.}

\item{sigma}{Dispersion parameter, where sigma > 0.}

\item{delta}{Proportion of extreme values, where 0 < delta < 1.}

\item{eta}{Inflation parameter, where eta > 1.}

\item{log}{Logical; if TRUE, probabilities are returned as log(p). Default is FALSE.}
}
\value{
The density of the contaminated beta-binomial distribution.
}
\description{
Density function of a contaminated mean-parameterized beta-binomial distribution.
}
\details{
The contaminated beta-binomial distribution with parameters mu (\eqn{\mu}), sigma (\eqn{\sigma}), delta (\eqn{\delta}), and eta (\eqn{\eta}) has a density given by:
\deqn{
f(x; m, \mu, \sigma, \delta, \eta) = (1 - \delta) \cdot \text{BetaBinom}(x; m, \mu, \sigma) + \delta \cdot \text{BetaBinom}(x; m, \mu, \eta \sigma)
}
where \eqn{\text{BetaBinom}(x; m, \mu, \sigma)} is the density of a beta-binomial distribution with:
- \eqn{n} = number of trials (\code{size}),
Constraints: \eqn{x = 0, 1, \dots, m}, \eqn{0 < \mu < 1}, \eqn{\sigma > 0}, \eqn{0 < \delta < 1}, \eqn{\eta > 1}.
}
