% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmbetabinom.R
\name{dmbetabinom}
\alias{dmbetabinom}
\title{Mean-Parameterized Beta-Binomial Density}
\usage{
dmbetabinom(x, size, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Vector of non-negative integer quantities.}

\item{size}{Number (or vector of length x) of trials (non-negative integer).}

\item{mu}{Analogue of the binomial probability of success parameter, where 0 < mu < 1.}

\item{sigma}{Dispersion parameter, where sigma > 0.}

\item{log}{Logical; if TRUE, probabilities are returned as log(p). Default is FALSE.}
}
\value{
The density of the beta-binomial distribution.
}
\description{
Density function of a mean-parameterized beta-binomial distribution.
}
\details{
The beta-binomial distribution with parameters mu (\eqn{\mu}) and sigma (\eqn{\sigma}) has a density given by:
\deqn{
f(x; n, \mu, \sigma) = \binom{m}{x} \frac{B(x + \frac{\mu}{\sigma}, m - x + \frac{1-\mu}{\sigma})}{B(\frac{\mu}{\sigma},\frac{1-\mu}{\sigma})}
}
where:
- \eqn{m} = number of trials (\code{size}),
- \eqn{B} is the beta function.
Constraints: \eqn{x = 0, 1, \dots, m}, \eqn{0 < \mu < 1}, \eqn{\sigma > 0}.
}
