% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.cmbb.R
\name{ml.cmbb}
\alias{ml.cmbb}
\title{Maximum Likelihood Estimation of Contaminated Beta-Binomial Regression}
\usage{
ml.cmbb(
  formula,
  sigma.formula = ~1,
  delta.formula = ~1,
  eta.formula = ~1,
  data,
  start = NULL,
  method = "BFGS",
  reltol = 1e-15,
  maxit = 1000,
  hessian = T
)
}
\arguments{
\item{formula}{An object of class 'formula': a symbolic description of the model for the mean parameter (mu).}

\item{sigma.formula}{A formula for the dispersion parameter (sigma). Defaults to ~1.}

\item{delta.formula}{A formula for the proportion of extreme values parameter (delta). Defaults to ~1.}

\item{eta.formula}{A formula for the inflation parameter (eta). Defaults to ~1.}

\item{data}{A mandatory data frame containing the variables in the model.}

\item{start}{Vector of initial values. If NULL, values are initialized using ml.mbb and with delta = 0.01 and eta = 1.1.}

\item{method}{Optimization method to be used. Default is "BFGS". Other options are "Nelder-Mead".}

\item{reltol}{Relative convergence tolerance in optimization. Defaults to 1e-15.}

\item{maxit}{The maximum number of iterations in optimization. Defaults to 1000.}

\item{hessian}{Logical; if TRUE, computes the Hessian matrix for standard errors. Default is TRUE.}

\item{verbose}{Logical; if TRUE, prints progress information. Default is FALSE.}
}
\value{
A list of elements:
   \item{results}{A data frame with parameter estimates, standard errors, t-values, and p-values (if hessian = TRUE).}
   \item{beta}{Maximum likelihood estimates of the regression coefficients for mu.}
   \item{alpha}{Maximum likelihood estimates of the regression coefficients for sigma.}
   \item{gamma}{Maximum likelihood estimates of the regression coefficients for delta.}
   \item{lambda}{Maximum likelihood estimates of the regression coefficients for eta.}
   \item{mu}{Fitted values of the mean parameter.}
   \item{sigma}{Fitted values of the dispersion parameter.}
   \item{delta}{Fitted values of the contamination proportion.}
   \item{eta}{Fitted values of the inflation parameter.}
   \item{X}{The design matrix for mu.}
   \item{U}{The design matrix for sigma.}
   \item{V}{The design matrix for delta.}
   \item{Z}{The design matrix for eta.}
   \item{y}{The response variable.}
   \item{loglike}{The log-likelihood value at convergence.}
   \item{AIC}{Akaike Information Criterion (AIC) for the fitted model.}
   \item{BIC}{Bayesian Information Criterion (BIC) for the fitted model.}
   \item{HIC}{Hannan-Quinn Information Criterion (HIC) for the fitted model.}
}
\description{
Maximum likelihood estimation of the contaminated mean-parameterized beta-binomial regression model via direct optimization.
}
\details{
The \code{ml.cmbb} function fits a contaminated beta-binomial regression model
}
