% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.mbb.R
\name{ml.mbb}
\alias{ml.mbb}
\title{Maximum Likelihood Estimation of Beta-Binomial Regression}
\usage{
ml.mbb(
  formula,
  sigma.formula = ~1,
  data,
  start = NULL,
  method = "BFGS",
  reltol = 1e-15,
  maxit = 1000,
  hessian = T
)
}
\arguments{
\item{formula}{An object of class 'formula': a symbolic description of the model to be fitted.}

\item{sigma.formula}{A formula for the dispersion parameter (sigma). Defaults to ~1.}

\item{data}{A mandatory data frame containing the variables in the model.}

\item{start}{Vector of initial values. If NULL, values produced by glm(family=Binomial) are used as initial values.}

\item{method}{Optimization method to be used. Default is "BFGS". Other options are "Nelder-Mead".}

\item{reltol}{Relative convergence tolerance in each optimization process. Defaults to 1e-15.}

\item{maxit}{The number of inner iterations in each optimization process. Defaults to 1000.}

\item{hessian}{Logical; if TRUE, computes the Hessian matrix for standard errors. Default is TRUE.}
}
\value{
A list of elements:
   \item{results}{A data frame with parameter estimates and standard errors.}
   \item{alpha}{Maximum likelihood estimate of the coefficients for sigma.}
   \item{beta}{Maximum likelihood estimates of the regression coefficients for mu.}
   \item{mu}{Fitted values of the mean parameter.}
   \item{sigma}{Fitted values of the dispersion parameter.}
   \item{X}{The design matrix for mu.}
   \item{U}{The design matrix for sigma.}
   \item{y}{The response variable.}
   \item{loglike}{The log-likelihood value at convergence.}
   \item{AIC}{Akaike Information Criterion (AIC) for the fitted model.}
   \item{BIC}{Bayesian Information Criterion (BIC) for the fitted model.}
   \item{HIC}{Hannan-Quinn Information Criterion (HIC) for the fitted model.}
}
\description{
Maximum likelihood estimation of the mean-parameterized beta-binomial regression model.
}
\details{
The \code{ml.mbb} function fits the mean-parameterized beta-binomial regression model.
}
