% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmbb_regression.R
\name{rcmbb_regression}
\alias{rcmbb_regression}
\title{Simulate Contaminated Beta-Binomial Regression Data}
\usage{
rcmbb_regression(
  n = 1000,
  size = 10,
  beta = c(1, 0.75),
  alpha = c(log(0.5)),
  gamma = c(log(0.05/(1 - 0.05))),
  lambda = c(log(1.2 - 1))
)
}
\arguments{
\item{n}{Number of observations to simulate. Default is 1000.}

\item{size}{Number of trials for the beta-binomial distribution. Default is 10.}

\item{beta}{Vector of regression coefficients for the mean parameter (mu). Default is c(1, 0.75).}

\item{alpha}{Vector of regression coefficients for the dispersion parameter (sigma). Default is c(log(0.5)).}

\item{gamma}{Vector of regression coefficients for the contamination proportion (delta). Default is c(log(0.05 / (1 - 0.05))).}

\item{lambda}{Vector of regression coefficients for the inflation parameter (eta). Default is c(log(1.2 - 1)).}
}
\value{
A data frame containing:
   \item{cbby}{The simulated contaminated beta-binomial responses.}
   \item{x0, x1, ...}{The design matrix columns for mu (including intercept).}
   \item{u0, u1, ...}{The design matrix columns for sigma (including intercept).}
   \item{v0, v1, ...}{The design matrix columns for delta (including intercept).}
   \item{z0, z1, ...}{The design matrix columns for eta (including intercept).}
}
\description{
Simulates data from a contaminated mean-parameterized beta-binomial regression model.
}
\details{
The \code{rcbb_regression} function simulates data from a contaminated beta-binomial regression model
The predictors \eqn{X}, \eqn{U}, \eqn{V}, and \eqn{Z} are generated as uniform random variables in [-2, 2].
}
