% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmbb_regression.R
\name{rmbb_regression}
\alias{rmbb_regression}
\title{Simulate Beta-Binomial Regression Data}
\usage{
rmbb_regression(n = 100, size = 10, beta = c(1, 0.75, -1.5), alpha = c(2, 5))
}
\arguments{
\item{n}{Number of observations to simulate. Default is 100.}

\item{size}{Number of trials for the beta-binomial distribution. Default is 10.}

\item{beta}{Vector of regression coefficients for the mean parameter (mu). Default is c(1, 0.75, -1.5).}

\item{alpha}{Vector of regression coefficients for the dispersion parameter (sigma). Default is c(2, 5).}
}
\value{
A data frame containing:
   \item{bby}{The simulated beta-binomial responses.}
   \item{x0, x1, ...,xm}{The design matrix columns for mu (including intercept).}
   \item{u0, u1, ...,um}{The design matrix columns for sigma (including intercept).}
}
\description{
Simulates data from a mean-parameterized beta-binomial regression model.
}
\details{
The \code{rbb_regression} function simulates data from a beta-binomial regression model where the response follows a beta-binomial distribution with mean \eqn{\mu} and dispersion \eqn{\sigma}.]
The predictors \eqn{X} and \eqn{U} are generated as uniform random variables in [-2, 2].
}
